new domain objects, better strategyes, handling, implem changes

+++ src/main/java/com/bigbank/dragons/api/dto/BatchStatsDto.java
@@ -3,7 +3,7 @@ package com.bigbank.dragons.api.dto;
 public record BatchStatsDto(
     int games,
     double averageScore,
-    int maxScore,
-    int minScore,
+    double maxScore,
+    double minScore,
     long gamesReachedTarget,
     double reachedTargetPercent) {}
